backend: improve admin interface and update requirements (#46)
Add admin interfaces for database models. May not be complete or fully functional, but does allow the user id on activities to be changed which was the motivation for this change.

django now defaults to the local settings module so setting the environment variable is no longer necessary for local development.